Stop using small integral types like `int8_t`

+++ src/ECS/TransformComponent.hpp
@@ -9,11 +9,11 @@ public:
 	Vector2D position;
 	Vector2D velocity;
 	
-	uint_fast16_t height = 92;
-	uint_fast16_t width = 66;
-	int_fast8_t scale = 1;
+	uint_fast32_t height = 92;
+	uint_fast32_t width = 66;
+	int_fast32_t scale = 1;
 
-	uint_fast8_t speed = 3;
+	uint_fast32_t speed = 3;
 
 	TransformComponent() {
 		position.Zero();
@@ -22,14 +22,14 @@ public:
 		position.x = x;
 		position.y = y;
 	}
-	TransformComponent(float x, float y, uint_fast16_t h, uint_fast16_t w, int_fast8_t sc) {
+	TransformComponent(float x, float y, uint_fast32_t h, uint_fast32_t w, int_fast32_t sc) {
 		position.x = x;
 		position.y = y;
 		height = h;
 		width = w;
 		scale = sc;
 	}
-	TransformComponent(int_fast8_t sc) {
+	TransformComponent(int_fast32_t sc) {
 		position.Zero();
 		scale = sc;
 	}
